ForecastData used as main state of the main screen @Path changed to @Query

+++ app/src/main/kotlin/com/alex/myweather/data/model/remote/Day.kt
@@ -4,24 +4,24 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Day(
-    val avghumidity: Int,
+    val avghumidity: Double,
     val avgtemp_c: Double,
     val avgtemp_f: Double,
-    val avgvis_km: Int,
-    val avgvis_miles: Int,
+    val avgvis_km: Double,
+    val avgvis_miles: Double,
     val condition: Condition,
-    val daily_chance_of_rain: Int,
-    val daily_chance_of_snow: Int,
-    val daily_will_it_rain: Int,
-    val daily_will_it_snow: Int,
+    val daily_chance_of_rain: Double,
+    val daily_chance_of_snow: Double,
+    val daily_will_it_rain: Double,
+    val daily_will_it_snow: Double,
     val maxtemp_c: Double,
     val maxtemp_f: Double,
     val maxwind_kph: Double,
     val maxwind_mph: Double,
     val mintemp_c: Double,
     val mintemp_f: Double,
-    val totalprecip_in: Int,
-    val totalprecip_mm: Int,
-    val totalsnow_cm: Int,
-    val uv: Int
+    val totalprecip_in: Double,
+    val totalprecip_mm: Double,
+    val totalsnow_cm: Double,
+    val uv: Double
 )
